Use a single file (index.d.ts) for typescript
Avoid issues with import / export and module resolution
