--- types/Options.d.ts
@@ -1,11 +1,0 @@
-export interface RequestOptions {
-  projectId?: number;
-}
-
-export interface PushpadOptions {
-  authToken: string;
-  projectId?: number;
-  baseUrl?: string;
-  fetch?: typeof fetch;
-  timeout?: number;
-}
