--- types/index.d.ts
@@ -1,12 +1,0 @@
-export * from './Notification';
-export * from './NotificationResource';
-export * from './Subscription';
-export * from './SubscriptionResource';
-export * from './Project';
-export * from './ProjectResource';
-export * from './Sender';
-export * from './SenderResource';
-export * from './Options';
-export * from './PushpadError';
-export { default } from './Pushpad';
-export { Pushpad } from './Pushpad';
